requirements.txt
alembic==1.4.3
appnope==0.1.0
attrs==20.2.0
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
decorator==4.4.2
Django==2.2
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-Testing==0.8.0
idna==2.10
importlib-metadata==2.0.0
iniconfig==1.1.1
ipdb==0.13.3
ipython==7.18.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
packaging==20.4
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.7
ptyprocess==0.6.0
py==1.9.0
Pygments==2.7.1
PyMySQL==0.10.1
pyparsing==2.4.7
pytest==6.1.2
pytest-html==2.1.1
pytest-metadata==1.10.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.20
sqlparse==0.3.1
toml==0.10.1
traitlets==5.0.4
urllib3==1.25.11
wcwidth==0.2.5
Werkzeug==1.0.1
zipp==3.4.0
marshmallow==3.11.1
